Abstracts test cases into a single class
