+++ integration-test/src/test/kotlin/io/github/pak3nuh/messaging/outbox/locking/MySqlLockTest.kt
@@ -1,65 +1,26 @@
 package io.github.pak3nuh.messaging.outbox.locking
 
 import io.github.pak3nuh.util.logging.KLoggerFactory
-import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.Test
 import org.testcontainers.containers.GenericContainer
 import org.testcontainers.containers.output.Slf4jLogConsumer
 import org.testcontainers.junit.jupiter.Container
 import org.testcontainers.junit.jupiter.Testcontainers
-import java.time.Duration
-import kotlin.concurrent.thread
 
 @Testcontainers
-class MySqlLockTest {
+class MySqlLockTest: AbstractSqlLockTest() {
 
-    @Test
-    fun `should obtain lock`() {
-        val lock = getLock("obtain-lock")
-        lock.close()
-    }
+    override val container: GenericContainer<*>
+        get() = Companion.container
 
-    @Test
-    fun `should wait lock release`() {
-        try {
-            Assertions.assertTimeoutPreemptively(Duration.ofSeconds(2)) {
-                val id = "lock-release"
-                val lock1 = getLock(id)
-                val lock2 = getLock(id)
-            }
-        } catch (error: AssertionError) {
-            val message = error.cause?.message
-            println(message)
-            Assertions.assertTrue(message?.contains("Execution timed out in thread") ?: error("no message"))
+    override val driverManagerProvider: DriverManagerProvider
+        get() {
+            val host = container.host
+            val port = container.firstMappedPort
+            return DriverManagerProvider("jdbc:mysql://$host:$port/database", "mysql", "mysql")
         }
-    }
-
-    @Test
-    fun `should not block different ids`() {
-        Assertions.assertTimeoutPreemptively(Duration.ofSeconds(2)) {
-            val lock1 = getLock("lock_id_1")
-            val lock2 = getLock("lock_id_2")
-        }
-    }
-
-    @Test
-    fun `should release lock on connection close`() {
-        Assertions.assertTimeoutPreemptively(Duration.ofSeconds(2)) {
-            val id = "release-on-connection-close"
-            val lock1 = getLock(id)
-            thread { lock1.close() }
-            val lock2 = getLock(id)
-        }
-    }
-
-    private fun getLock(id: String): Lock {
-        return sut.tryLock(id, Duration.ZERO) ?: error("Lock not acquired")
-    }
 
     companion object {
-
-        private val logger = KLoggerFactory.getLogger<MySqlLockTest>()
 
         @JvmStatic
         @Container
@@ -73,19 +34,10 @@ class MySqlLockTest {
             .withExposedPorts(3306)
             .withFileSystemBind("src/test/resources/init-mysql.sql", "/docker-entrypoint-initdb.d/init.sql")
 
-        lateinit var sut: LockFactory
-
         @JvmStatic
         @BeforeAll
         fun initContainer() {
-            container.followOutput(Slf4jLogConsumer(logger))
-
-            val host = container.host
-            val port = container.firstMappedPort
-            sut = LockFactoryImpl(SqlProvider(
-                "application_locks",
-                DriverManagerProvider("jdbc:mysql://$host:$port/database", "mysql", "mysql")
-            ))
+            container.followOutput(Slf4jLogConsumer(KLoggerFactory.getLogger<MySqlLockTest>()))
         }
     }
 }
